RadiationWatch.(cpp|h): float -> double
Lets use double on platforms that support double (8-byte). Arduino on
ATMEGA based boards use only 4-byte (=float) when using double [1], so
no change for this platform.

[1] https://www.arduino.cc/en/Reference/Double

+++ RadiationWatch.cpp
@@ -156,22 +156,22 @@ unsigned long RadiationWatch::radiationCount() {
   return _count;
 }
 
-float RadiationWatch::cpm()
+double RadiationWatch::cpm()
 {
   // cpm = uSv x alpha
-  float min = integrationTime() / 60000.0;
+  double min = integrationTime() / 60000.0;
   return (min > 0) ? radiationCount() / min : 0;
 }
 
-static const float kAlpha = 53.032;
+static const double kAlpha = 53.032;
 
-float RadiationWatch::uSvh()
+double RadiationWatch::uSvh()
 {
   return cpm() / kAlpha;
 }
 
-float RadiationWatch::uSvhError()
+double RadiationWatch::uSvhError()
 {
-  float min = integrationTime() / 60000.0;
+  double min = integrationTime() / 60000.0;
   return (min > 0) ? sqrt(radiationCount()) / min / kAlpha : 0;
 }
